+++ CHANGELOG.md
@@ -5,6 +5,18 @@
 
 - [usd#2652](https://github.com/Autodesk/arnold-usd/issues/2652) - Add setting for accelerated viewport
 
+## [7.5.4.0] (Unreleased)
+
+### Features
+
+- [usd#2705](https://github.com/Autodesk/arnold-usd/issues/2705) - Bind volume shaders on step-sized meshes
+- [usd#2719](https://github.com/Autodesk/arnold-usd/issues/2719) - Add support for resumable rendering in the render delegate.
+
+### Bug fixes
+
+- [usd#2722](https://github.com/Autodesk/arnold-usd/issues/2722) - Fix MSVC Linking error in NormalsPruningDataSource scene index.
+
+
 ## [7.5.3.1] (Unreleased)
 
 ### Bug fixes
@@ -12,7 +24,7 @@
 - [usd#2699](https://github.com/Autodesk/arnold-usd/issues/2699) - Fix primvars:crypto_asset being ignored on instanceable references
 
 
-## [7.5.3.0] (Unreleased)
+## [7.5.3.0] 2026-07-22
 
 ### Features